# back/src/api/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Clothing": {
				"properties": {
					"id": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"imageUrl",
					"description",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Partial_Clothing_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ai-closet-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/clothing": {
			"get": {
				"operationId": "GetClothes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Clothing"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Clothing"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "color",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateClothing",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Clothing"
								}
							}
						}
					}
				},
				"tags": [
					"Clothing"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Clothing_"
							}
						}
					}
				}
			}
		},
		"/clothing/{id}": {
			"get": {
				"operationId": "GetClothingById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Clothing"
								}
							}
						}
					}
				},
				"tags": [
					"Clothing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateClothing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Clothing"
								}
							}
						}
					}
				},
				"tags": [
					"Clothing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Clothing_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteClothing",
				"responses": {
					"204": {
						"description": "Deleted"
					}
				},
				"tags": [
					"Clothing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}